// src/models/status.ts
export interface Status {
    id: number;
    health: number;
    maxHealth: number;
    healthRegen: number;
    mana: number;
    maxMana: number;
    manaRegen: number;
    strength: number;
    dexterity: number;
    wisdom: number;
    perception: number;
    damage: number;
    maxDamage: number;
    attackSpeed: number;
    meleeDefence: number;
    magicDefence: number;
    evasion: number;
    criticalDamage: number;
}